install: Split off baseline module
Since we added `install-to-filesystem`, the install code needs
a refactoring where we more cleanly split up our "default"
installer which in turn passes its logic to the filesystem
side.
