+++ lib/src/install.rs
@@ -1,9 +1,15 @@
-use std::borrow::Cow;
-use std::fmt::Display;
+//! # Writing a container to a block device in a bootable way
+//!
+//! This module supports installing a bootc-compatible image to
+//! a block device directly via the `install` verb, or to an externally
+//! set up filesystem via `install-to-filesystem`.
+
+// This sub-module is the "basic" installer that handles creating basic block device
+// and filesystem setup.
+mod baseline;
+
 use std::io::BufWriter;
 use std::io::Write;
-use std::process::Command;
-use std::process::Stdio;
 use std::str::FromStr;
 use std::sync::Arc;
 
@@ -15,7 +21,7 @@ use cap_std::fs::Dir;
 use cap_std_ext::cap_std;
 use cap_std_ext::prelude::CapStdExtDirExt;
 use cap_std_ext::rustix::fs::MetadataExt;
-use clap::ArgEnum;
+
 use fn_error_context::context;
 use ostree::gio;
 use ostree_ext::container as ostree_container;
@@ -24,6 +30,7 @@ use ostree_ext::ostree;
 use ostree_ext::prelude::Cast;
 use serde::{Deserialize, Serialize};
 
+use self::baseline::InstallBlockDeviceOpts;
 use crate::lsm::lsm_label;
 use crate::task::Task;
 use crate::utils::run_in_host_mountns;
@@ -37,56 +44,8 @@ const RUN_BOOTC: &str = "/run/bootc";
 /// This is an ext4 special directory we need to ignore.
 const LOST_AND_FOUND: &str = "lost+found";
 
-#[derive(clap::ValueEnum, Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub(crate) enum BlockSetup {
-    Direct,
-    Tpm2Luks,
-}
-
-impl Default for BlockSetup {
-    fn default() -> Self {
-        Self::Direct
-    }
-}
-
-#[derive(clap::ValueEnum, Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) enum Filesystem {
-    Xfs,
-    Ext4,
-    Btrfs,
-}
-
-impl Default for Filesystem {
-    fn default() -> Self {
-        // Obviously this should be configurable.
-        Self::Xfs
-    }
-}
-
-impl Display for Filesystem {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.to_possible_value().unwrap().get_name().fmt(f)
-    }
-}
-
 /// Kernel argument used to specify we want the rootfs mounted read-write by default
 const RW_KARG: &str = "rw";
-
-const BOOTPN: u32 = 3;
-// This ensures we end up under 512 to be small-sized.
-const BOOTPN_SIZE_MB: u32 = 510;
-const ROOTPN: u32 = 4;
-#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
-const EFIPN: u32 = 2;
-#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
-const EFIPN_SIZE_MB: u32 = 512;
-#[cfg(target_arch = "aarch64")]
-const RESERVEDPN: u32 = 1;
-#[cfg(target_arch = "ppc64")]
-const PREPPN: u32 = 1;
-#[cfg(target_arch = "ppc64")]
-const RESERVEDPN: u32 = 1;
 
 #[derive(clap::Args, Debug, Clone, Serialize, Deserialize)]
 pub(crate) struct InstallTargetOpts {
@@ -139,38 +98,6 @@ pub(crate) struct InstallConfigOpts {
     #[clap(long)]
     /// Add a kernel argument
     karg: Option<Vec<String>>,
-}
-
-/// Options for installing to a block device
-#[derive(Debug, Clone, clap::Args, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub(crate) struct InstallBlockDeviceOpts {
-    /// Target block device for installation.  The entire device will be wiped.
-    pub(crate) device: Utf8PathBuf,
-
-    /// Automatically wipe all existing data on device
-    #[clap(long)]
-    #[serde(default)]
-    pub(crate) wipe: bool,
-
-    /// Target root block device setup.
-    ///
-    /// direct: Filesystem written directly to block device
-    /// tpm2-luks: Bind unlock of filesystem to presence of the default tpm2 device.
-    #[clap(long, value_enum, default_value_t)]
-    #[serde(default)]
-    pub(crate) block_setup: BlockSetup,
-
-    /// Target root filesystem type.
-    #[clap(long, value_enum, default_value_t)]
-    #[serde(default)]
-    pub(crate) filesystem: Filesystem,
-
-    /// Size of the root partition (default specifier: M).  Allowed specifiers: M (mebibytes), G (gibibytes), T (tebibytes).
-    ///
-    /// By default, all remaining space on the disk will be used.
-    #[clap(long)]
-    pub(crate) root_size: Option<String>,
 }
 
 /// Perform an installation to a block device.
@@ -232,7 +159,7 @@ pub(crate) struct InstallToFilesystemOpts {
 }
 
 // Shared read-only global state
-struct State {
+pub(crate) struct State {
     /// Image reference we'll pull from (today always containers-storage: type)
     source_imageref: ostree_container::ImageReference,
     /// The digest to use for pulls
@@ -334,61 +261,6 @@ impl FromStr for MountSpec {
             options,
         })
     }
-}
-
-fn sgdisk_partition(
-    sgdisk: &mut Command,
-    n: u32,
-    part: impl AsRef<str>,
-    name: impl AsRef<str>,
-    typecode: Option<&str>,
-) {
-    sgdisk.arg("-n");
-    sgdisk.arg(format!("{n}:{}", part.as_ref()));
-    sgdisk.arg("-c");
-    sgdisk.arg(format!("{n}:{}", name.as_ref()));
-    if let Some(typecode) = typecode {
-        sgdisk.arg("-t");
-        sgdisk.arg(format!("{n}:{typecode}"));
-    }
-}
-
-fn mkfs<'a>(
-    dev: &str,
-    fs: Filesystem,
-    label: Option<&'_ str>,
-    opts: impl IntoIterator<Item = &'a str>,
-) -> Result<uuid::Uuid> {
-    let u = uuid::Uuid::new_v4();
-    let mut t = Task::new("Creating filesystem", format!("mkfs.{fs}"));
-    match fs {
-        Filesystem::Xfs => {
-            t.cmd.arg("-m");
-            t.cmd.arg(format!("uuid={u}"));
-        }
-        Filesystem::Btrfs | Filesystem::Ext4 => {
-            t.cmd.arg("-U");
-            t.cmd.arg(u.to_string());
-        }
-    };
-    // Today all the above mkfs commands take -L
-    if let Some(label) = label {
-        t.cmd.args(["-L", label]);
-    }
-    t.cmd.args(opts);
-    t.cmd.arg(dev);
-    // All the mkfs commands are unnecessarily noisy by default
-    t.cmd.stdout(Stdio::null());
-    t.run()?;
-    Ok(u)
-}
-
-fn mount(dev: &str, target: &Utf8Path) -> Result<()> {
-    Task::new_and_run(
-        format!("Mounting {target}"),
-        "mount",
-        [dev, target.as_str()],
-    )
 }
 
 fn bind_mount_from_host(src: impl AsRef<Utf8Path>, dest: impl AsRef<Utf8Path>) -> Result<()> {
@@ -601,7 +473,7 @@ fn skopeo_supports_containers_storage() -> Result<bool> {
     Ok(major > 1 || minor > 10)
 }
 
-struct RootSetup {
+pub(crate) struct RootSetup {
     device: Utf8PathBuf,
     rootfs: Utf8PathBuf,
     rootfs_fd: Dir,
@@ -620,178 +492,6 @@ impl RootSetup {
     fn get_boot_uuid(&self) -> Result<&str> {
         require_boot_uuid(&self.boot)
     }
-}
-
-#[context("Creating rootfs")]
-fn install_create_rootfs(state: &State, opts: InstallBlockDeviceOpts) -> Result<RootSetup> {
-    // Verify that the target is empty (if not already wiped in particular, but it's
-    // also good to verify that the wipe worked)
-    let device = crate::blockdev::list_dev(&opts.device)?;
-
-    // Handle wiping any existing data
-    if opts.wipe {
-        let dev = &opts.device;
-        for child in device.children.iter().flatten() {
-            let child = child.path();
-            println!("Wiping {child}");
-            crate::blockdev::wipefs(Utf8Path::new(&child))?;
-        }
-        println!("Wiping {dev}");
-        crate::blockdev::wipefs(dev)?;
-    } else if device.has_children() {
-        anyhow::bail!(
-            "Detected existing partitions on {}; use e.g. `wipefs` if you intend to overwrite",
-            opts.device
-        );
-    }
-
-    // Now at this point, our /dev is a stale snapshot because we don't have udev running.
-    // So from hereon after, we prefix devices with our temporary devtmpfs mount.
-    let reldevice = opts
-        .device
-        .strip_prefix("/dev/")
-        .context("Absolute device path in /dev/ required")?;
-    let device = state.devdir.join(reldevice);
-
-    let root_size = opts
-        .root_size
-        .as_deref()
-        .map(crate::blockdev::parse_size_mib)
-        .transpose()
-        .context("Parsing root size")?;
-
-    // Create a temporary directory to use for mount points.  Note that we're
-    // in a mount namespace, so these should not be visible on the host.
-    let rootfs = state.mntdir.join("rootfs");
-    std::fs::create_dir_all(&rootfs)?;
-    let bootfs = state.mntdir.join("boot");
-    std::fs::create_dir_all(bootfs)?;
-
-    // Run sgdisk to create partitions.
-    let mut sgdisk = Task::new("Initializing partitions", "sgdisk");
-    // sgdisk is too verbose
-    sgdisk.cmd.stdout(Stdio::null());
-    sgdisk.cmd.arg("-Z");
-    sgdisk.cmd.arg(&device);
-    sgdisk.cmd.args(["-U", "R"]);
-    #[allow(unused_assignments)]
-    if cfg!(target_arch = "x86_64") {
-        // BIOS-BOOT
-        sgdisk_partition(
-            &mut sgdisk.cmd,
-            1,
-            "0:+1M",
-            "BIOS-BOOT",
-            Some("21686148-6449-6E6F-744E-656564454649"),
-        );
-    } else if cfg!(target_arch = "aarch64") {
-        // reserved
-        sgdisk_partition(
-            &mut sgdisk.cmd,
-            1,
-            "0:+1M",
-            "reserved",
-            Some("8DA63339-0007-60C0-C436-083AC8230908"),
-        );
-    } else {
-        anyhow::bail!("Unsupported architecture: {}", std::env::consts::ARCH);
-    }
-
-    let espdev = if cfg!(any(target_arch = "x86_64", target_arch = "aarch64")) {
-        sgdisk_partition(
-            &mut sgdisk.cmd,
-            EFIPN,
-            format!("0:+{EFIPN_SIZE_MB}M"),
-            "EFI-SYSTEM",
-            Some("C12A7328-F81F-11D2-BA4B-00A0C93EC93B"),
-        );
-        Some(format!("{device}{EFIPN}"))
-    } else {
-        None
-    };
-
-    sgdisk_partition(
-        &mut sgdisk.cmd,
-        BOOTPN,
-        format!("0:+{BOOTPN_SIZE_MB}M"),
-        "boot",
-        None,
-    );
-    let root_size = root_size
-        .map(|v| Cow::Owned(format!("0:{v}M")))
-        .unwrap_or_else(|| Cow::Borrowed("0:0"));
-    sgdisk_partition(
-        &mut sgdisk.cmd,
-        ROOTPN,
-        root_size,
-        "root",
-        Some("0FC63DAF-8483-4772-8E79-3D69D8477DE4"),
-    );
-    sgdisk.run()?;
-
-    // Reread the partition table
-    {
-        let mut f = std::fs::OpenOptions::new()
-            .write(true)
-            .open(&device)
-            .with_context(|| format!("opening {device}"))?;
-        crate::blockdev::reread_partition_table(&mut f, true)
-            .context("Rereading partition table")?;
-    }
-
-    crate::blockdev::udev_settle()?;
-
-    match opts.block_setup {
-        BlockSetup::Direct => {}
-        // TODO
-        BlockSetup::Tpm2Luks => anyhow::bail!("tpm2-luks is not implemented yet"),
-    }
-
-    // TODO: make this configurable
-    let bootfs_type = Filesystem::Ext4;
-
-    // Initialize the /boot filesystem
-    let bootdev = &format!("{device}{BOOTPN}");
-    let boot_uuid = mkfs(bootdev, bootfs_type, Some("boot"), []).context("Initializing /boot")?;
-
-    // Initialize rootfs
-    let rootdev = &format!("{device}{ROOTPN}");
-    let root_uuid = mkfs(rootdev, opts.filesystem, Some("root"), [])?;
-    let rootarg = format!("root=UUID={root_uuid}");
-    let bootsrc = format!("UUID={boot_uuid}");
-    let bootarg = format!("boot={bootsrc}");
-    let boot = MountSpec::new(bootsrc.as_str(), "/boot");
-    let kargs = vec![rootarg, RW_KARG.to_string(), bootarg];
-
-    mount(rootdev, &rootfs)?;
-    lsm_label(&rootfs, "/".into(), false)?;
-    let rootfs_fd = Dir::open_ambient_dir(&rootfs, cap_std::ambient_authority())?;
-    let bootfs = rootfs.join("boot");
-    std::fs::create_dir(&bootfs).context("Creating /boot")?;
-    // The underlying directory on the root should be labeled
-    lsm_label(&bootfs, "/boot".into(), false)?;
-    mount(bootdev, &bootfs)?;
-    // And we want to label the root mount of /boot
-    lsm_label(&bootfs, "/boot".into(), false)?;
-
-    // Create the EFI system partition, if applicable
-    if let Some(espdev) = espdev {
-        Task::new("Creating ESP filesystem", "mkfs.fat")
-            .args([espdev.as_str(), "-n", "EFI-SYSTEM"])
-            .quiet_output()
-            .run()?;
-        let efifs_path = bootfs.join(crate::bootloader::EFI_DIR);
-        std::fs::create_dir(&efifs_path).context("Creating efi dir")?;
-        mount(&espdev, &efifs_path)?;
-    }
-
-    Ok(RootSetup {
-        device,
-        rootfs,
-        rootfs_fd,
-        boot,
-        kargs,
-    })
 }
 
 pub(crate) struct SourceData {
@@ -1024,8 +724,9 @@ pub(crate) async fn install(opts: InstallOpts) -> Result<()> {
 
     // This is all blocking stuff
     let mut rootfs = {
-        let state = state.clone();
-        tokio::task::spawn_blocking(move || install_create_rootfs(&state, block_opts)).await??
+        let state = Arc::clone(&state);
+        tokio::task::spawn_blocking(move || baseline::install_create_rootfs(&state, block_opts))
+            .await??
     };
 
     install_to_filesystem_impl(&state, &mut rootfs).await?;
